Initial version (#1)
The new version of webgear that exposes arrow based APIs. There are a ton of changes here from the previous version that this can be considered a new implementation for all practical purposes (except the concept of traits).

The documentation including haddock are not updated in this PR. I'll make another pass at it.